implemented integer key lookups rather than comparign by string. the QuestionMaster structure is great, but we have a bottleneck when checking the answer, since it's currently relying on string lookups in the main branch.
added Question class. ideally, we shouldn't rely on this. it is primarily an easy helper/communication object between other classes.
that being said, it might be good to have Questions use their own memory in addition to the structure's they're sitting in, for verification purposes.

either way, the above change allows the MainActivity class to know which question corresponds to which index value.

changed main activity class to use Question class

renamed and refactored the QuestionAndAnswerList to be "AnswerKey"

AnswerKey and QuestionKey now both point a given integer to a specifc corresponding value in each map, rather than question strings to answers.

I have tested absolutely none of this.